Initial commit
Repo already contains a LICENSE and README, with a simple script to
install the .emacs.d directory from cfg.

The meatiest part of the repo for now is my init.el, which already has
some configuration. Some nice packages are also now installed.

This repo was even checked in from Emacs via Magit, with the README
edited with markdown-mode and viewed with gh-md.

And so, my Emacs journey begins.

...this is fun. I should start moving to descriptive commit messages
for my other config repos.