Add project-specific packages
Projectile is a popular project management package, so I got it to
play around with it.

Indium and js2-mode seem to be the de-facto packages for Javascript
development, so I also got those. I think I'll get to use Emacs more
if I actually work on web-dev projects with it, instead of trying to
dabble solely for Lisp.