Restore original C-x C-f in Dired
I have a global-set-key for C-x C-f to run helm-find-files, so I can
make use of Helm's narrowing buffer for opening files.

However, the original find-file serves an important purpose in Dired:
quickly creating a new file in the current directory.

So, I simply reset C-x C-f in Dired mode with a local-set-key.