Use SLIME REPL instead of inferior buffer
I read on SLIME's info page that there exists a REPL that's better
integrated with Emacs than the usual REPL that comes with the inferior
Lisp buffer.

It's a contrib package, so I have to explicitly enable it.

Additionally, I removed an extraneous line where I include a
system-owned directory to my load-path (because I used to have SLIME
installed not by Emacs' packager, but by pacman).