Overhaul init.el
I'm returning to Emacs because I wanna give org another fair
shake. I'm yet again trying to resist going for evil, because some
quick tests show that I won't be able to use the keybinds listed in
the manual (e.g. M-RET and that sort of thing).

I'm coming at this from the perspective of only needing to edit Org
and Elisp files, so I'm not trying to set up an entire config for
development. Instead, I'm just setting up Org and Magit, then a bunch
of niceties like fuzzy finding and theming.